+++ Runtime/Scripts/Jobs.cs
@@ -1567,7 +1567,7 @@ namespace GLTFast.Jobs {
     }
 
     [BurstCompile]
-    public unsafe struct ConvertNormalsUInt16ToFloatInterleavedNormalizedJob : 
+    public unsafe struct ConvertPositionsUInt16ToFloatInterleavedNormalizedJob : 
 #if UNITY_JOBS
         IJobParallelForBatch
 #else
@@ -1599,7 +1599,7 @@ namespace GLTFast.Jobs {
                     off[1] / (float) ushort.MaxValue,
                     off[2] / (float) ushort.MaxValue
                 );
-                *resultV = normalize(tmp);
+                *resultV = tmp;
                 resultV = (float3*)((byte*)resultV + outputByteStride);
                 off = (ushort*)((byte*)off + inputByteStride);
             }
@@ -1608,12 +1608,11 @@ namespace GLTFast.Jobs {
         public void Execute(int i) {
             var resultV = (float3*) (((byte*)result) + (i*outputByteStride));
             var off = (ushort*) (input + (inputByteStride*i));
-            var tmp = new float3(
+            *resultV = new float3(
                 -(off[0] / (float) ushort.MaxValue),
                 off[1] / (float) ushort.MaxValue,
                 off[2] / (float) ushort.MaxValue
             );
-            *resultV = normalize(tmp);
         }
 #endif
     }
@@ -1662,6 +1661,65 @@ namespace GLTFast.Jobs {
 #endif
     }
 
+    /// <summary>
+    /// General purpose (position / morph target delta normal)
+    /// Result is not normalized (scaled to unit length)
+    /// </summary>
+    [BurstCompile]
+    public unsafe struct ConvertVector3Int16ToFloatInterleavedNormalizedJob : 
+#if UNITY_JOBS
+        IJobParallelForBatch
+#else
+        IJobParallelFor
+#endif
+    {
+
+        [ReadOnly]
+        public int inputByteStride;
+
+        [ReadOnly]
+        [NativeDisableUnsafePtrRestriction]
+        public byte* input;
+
+        [ReadOnly]
+        public int outputByteStride;
+
+        [ReadOnly]
+        [NativeDisableUnsafePtrRestriction]
+        public float3* result;
+
+#if UNITY_JOBS
+        public void Execute(int i, int count) {
+            var resultV = (float3*) ((byte*)result + i*outputByteStride);
+            var off = (short*) (input + i*inputByteStride);
+            
+            for (var x = 0; x < count; x++) {
+                var tmp = new float3(off[0], off[1], off[2]) / short.MaxValue;
+                var tmp2 = max(tmp, -1f);
+                tmp2.x *= -1;
+                *resultV = tmp2;
+                
+                resultV = (float3*)((byte*)resultV + outputByteStride);
+                off = (short*)((byte*)off + inputByteStride);
+            }
+        }
+#else
+        public void Execute(int i) {
+            var resultV = (float3*) (((byte*)result) + (i*outputByteStride));
+            var off = (short*) (input + (i*inputByteStride));
+
+            var tmp = new float3(off[0], off[1], off[2]) / short.MaxValue;
+            var tmp2 = max(tmp, -1f);
+            tmp2.x *= -1;
+            *resultV = tmp2;
+        }
+#endif
+    }
+
+    /// <summary>
+    /// Normal conversion
+    /// Result is normalized (scaled to unit length)
+    /// </summary>
     [BurstCompile]
     public unsafe struct ConvertNormalsInt16ToFloatInterleavedNormalizedJob : 
 #if UNITY_JOBS
@@ -1756,6 +1814,65 @@ namespace GLTFast.Jobs {
 #endif
     }
 
+    /// <summary>
+    /// General purpose conversion (positions or morph target delta normals)
+    /// Result is not normalized (scaled to unit length)
+    /// </summary>
+    [BurstCompile]
+    public unsafe struct ConvertVector3Int8ToFloatInterleavedNormalizedJob : 
+#if UNITY_JOBS
+        IJobParallelForBatch
+#else
+        IJobParallelFor
+#endif
+    {
+
+        [ReadOnly]
+        public int inputByteStride;
+
+        [ReadOnly]
+        [NativeDisableUnsafePtrRestriction]
+        public sbyte* input;
+
+        [ReadOnly]
+        public int outputByteStride;
+
+        [ReadOnly]
+        [NativeDisableUnsafePtrRestriction]
+        public float3* result;
+
+#if UNITY_JOBS
+        public void Execute(int i, int count) {
+            var resultV = (float3*) ((byte*)result + i*outputByteStride);
+            var off = input + i*inputByteStride;
+            
+            for (var x = 0; x < count; x++) {
+                var tmp = new float3(off[0], off[1], off[2]) / 127f;
+                var tmp2 = max(tmp, -1);
+                tmp2.x *= -1;
+                *resultV = tmp2;
+                
+                resultV = (float3*)((byte*)resultV + outputByteStride);
+                off += inputByteStride;
+            }
+        }
+#else
+        public void Execute(int i) {
+            var resultV = (float3*) (((byte*)result) + (i*outputByteStride));
+            var off = input + (inputByteStride*i);
+
+            var tmp = new float3(off[0], off[1], off[2]) / 127f;
+            var tmp2 = max(tmp, -1);
+            tmp2.x *= -1;
+            *resultV = tmp2;
+        }
+#endif
+    }
+
+    /// <summary>
+    /// Normal conversion
+    /// Result is normalized (scaled to unit length)
+    /// </summary>
     [BurstCompile]
     public unsafe struct ConvertNormalsInt8ToFloatInterleavedNormalizedJob : 
 #if UNITY_JOBS
@@ -1851,7 +1968,7 @@ namespace GLTFast.Jobs {
     }
 
     [BurstCompile]
-    public unsafe struct ConvertNormalsUInt8ToFloatInterleavedNormalizedJob : 
+    public unsafe struct ConvertPositionsUInt8ToFloatInterleavedNormalizedJob : 
 #if UNITY_JOBS
         IJobParallelForBatch
 #else
@@ -1879,12 +1996,11 @@ namespace GLTFast.Jobs {
             var off = input + i*inputByteStride;
             
             for (var x = 0; x < count; x++) {
-                var tmp = new float3(
+                *resultV = new float3(
                     -(off[0] / 255f),
                     off[1] / 255f,
                     off[2] / 255f
                 );
-                *resultV = normalize(tmp);
                 resultV = (float3*)((byte*)resultV + outputByteStride);
                 off += inputByteStride;
             }
@@ -1893,12 +2009,11 @@ namespace GLTFast.Jobs {
         public void Execute(int i) {
             var resultV = (float3*) (((byte*)result) + (i*outputByteStride));
             var off = input + (i*inputByteStride);
-            var tmp = new float3(
+            *resultV = new float3(
                 -(off[0] / 255f),
                 off[1] / 255f,
                 off[2] / 255f
             );
-            *resultV = normalize(tmp);
         }
 #endif
     }